\name{vtpnet-package}
\alias{vtpnet-package}
\alias{MdbSearchToDisk}
\alias{filterGWASbyMap}
\alias{getMtags}
\alias{getOneHits}
\alias{gwMatchPWM}
\alias{matrc}
\alias{rng2facHits}
\alias{runfimo}
\alias{setupGWPM}
\alias{cancerMap}
\alias{maurGWAS}
\alias{pax4}
\alias{pax4_75}
\alias{pax4_85}
\alias{vtpnet}
\docType{package}
\title{
variant-transcription factor-phenotype networks
}
\description{
variant-transcription factor-phenotype networks, inspired
        by Maurano et al., Science (2012), PMID 22955828
}
\details{
\tabular{ll}{
Package: \tab vtpnet\cr
Version: \tab 0.5.3\cr
Suggests: \tab MotifDb, VariantAnnotation, Rgraphviz\cr
Depends: \tab R (>= 3.0.0), graph, GenomicRanges, gwascat,
doParallel, foreach\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab yes\cr
biocViews: \tab Network\cr
Built: \tab R 3.1.1; ; 2014-10-08 14:42:02 UTC; unix\cr
}

Not for release at this time (Oct 2014).

}
\author{
VJ Carey <stvjc@channing.harvard.edu>

Maintainer: VJ Carey <stvjc@channing.harvard.edu>

}
\references{

}

\keyword{ package }
\seealso{
}
\examples{
}
